LPD-80239 Simplify the DropdownColorPicker component and change the name to TokenButton, which is now only a trigger and does not contain a dropdown
